extract code out to utils, hooks, components
