+++ apps/web/src/pages/SearchRequests/SearchPage/searchPageOperations.graphql
@@ -54,6 +54,7 @@ query getSearchFormDataAcrossAllPools {
   publishedPools {
     id
     owner {
+      id
       email
       firstName
       lastName
